requirements.txt
fastapi
functions-framework
tensorflow
google-cloud-storage
